[example-code] - LZ77: Decompression algorithm implemented. * Dictionary size limited to half the size of short.

+++ trunk/example-code/examples/LZ77/LZ77.h
@@ -6,13 +6,13 @@ using std::vector;
 
 struct EncodingPair
 {
-	int Reference;
+	short int Reference;
 	char Character;
 };
 
 struct DictionaryEntry
 {
-	int Reference;
+	short int Reference;
 	vector<char> Phrase;
 	EncodingPair Encoding;
 };
